server/requirements.txt
# Core dependencies
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
opencv-python>=4.4.0
Pillow>=8.0.0

# Server dependencies
flask>=2.0.0
flask-cors>=3.0.10
python-dotenv>=0.19.0
requests>=2.26.0

# Only needed for development
pytest>=6.0.0
